Update production.js
Swapped positions of the param definitions for the background color and assetPaths in the Umbra class to conform to their order in the constructor. As it is, the argument order in the documentation is wrong and always throws the assetPaths array error if the instance is constructed with more than the first arguments.